[1.3.2] Fixed craftbukkit compatibility (#562)
