+++ src/main/java/io/github/yannici/bedwars/Listener/PlayerListener.java
@@ -44,7 +44,6 @@ import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.SkullMeta;
 import org.bukkit.scheduler.BukkitRunnable;
-import org.spigotmc.event.player.PlayerSpawnLocationEvent;
 
 import io.github.yannici.bedwars.ChatWriter;
 import io.github.yannici.bedwars.Main;
@@ -103,22 +102,6 @@ public class PlayerListener extends BaseListener {
         }.runTaskLater(Main.getInstance(), 5L);
       }
 
-    }
-  }
-
-  @EventHandler(priority = EventPriority.HIGH)
-  public void onPlayerSpawnLocation(PlayerSpawnLocationEvent event) {
-    if (Main.getInstance().isBungee()) {
-      Player player = event.getPlayer();
-
-      ArrayList<Game> games = Main.getInstance().getGameManager().getGames();
-      if (games.size() == 0) {
-        return;
-      }
-
-      Game firstGame = games.get(0);
-
-      event.setSpawnLocation(firstGame.getPlayerTeleportLocation(player));
     }
   }
